Added kind/prelude
This is really exciting! I've added the first libary file:
`src/lib/prelude.kind`.

This library file is special in that the functions declared in it are
automatically imported. Currently that's only the case when calling
`kind:load`, but that still allows the tests for codegen, typegen and
the scanner to have this behaviour.

Currently, the prelude only declares the function `match`, which I had
originally implemented in the compiler as a special AST term, but
realised I could just implement as a function instead, and quite easily
too:

```
def match a f -> f(a)
```

I realise as I've changed codegen tests that this module is really
rather outdated and insufficient. It doesn't properly handle pattern
matching and falls short on quite a few other counts too. That's
something I'll need to address.

First on my list though is to make it possible to export `True` and
`False`. Currently I can export `Boolean` and import `Boolean/_` in a
module to have this effect, but I'd like for a module to be able to
specify subtypes that are imported with a wildcard import of the module
itself.