Create type module for compiled type members
We want to be able to import types and type members from a compiled (and
uncompiled) module.

There's several ways that it could be implemented. As far as I can see,
creating separate modules for each top-level type which contains
functions for sub-types is the easiest way to go about it.

This way, we don't have to encode separate state in a special function
in the module and we can use the existing import system for types as
well as functions.

For this change to come about, a few different pieces needed to be in
place in the typegen module to make sure that for every module
declaration in an AST and for every top level type, a new module is
generated with the appropriate number of exported functions.

Next step is to allow modules to export type members like `True` and
`False` on their own, so a module can import them without having to
write `Boolean/True` and `Boolean/False`.