Wire type modules through to code gen
If we export a type from a module, the compiled beam file must contain the type
function. For this to be the case we've wired the forms created in the
type generation through to the code generation.

If we import and use a type function to create other types, it must have
been compiled prior to this point in the compilation. For this reason we
need to sort the modules based on their dependencies. I've used erlangs
bigraph module for this purpose. It conveniently exposes a topological
sort that we can use to convert the dependency graph to a sorted list of
modules.

Thrown in with these changes are some testing utilities and a few bug
fixes in typegen.