Shifts responsibility of parsing std lib to parser
Before this change, the `kind` module would insert an extra line in the
source text adding an import of `kind/prelude/_`. Because this happened
before parsing, it meant that line numbers in errors were wrong.

Instead I'm adding the extra imports in the parsing module and adding a
couple of options to manage whether they are imported or not:

`add_kind_libraries`: Adds all `kind` files under `src/lib`
`import_prelude`: Imports the prelude in to all files (except the
prelude itself).