Fixed broken argument count in typegen test
A typegen test was broken with an error saying the number of arguments
for a call to the type function `Option` was wrong. The `collect_args`
function in `typegen` looked like it was doing the right thing when
debugging.

The issue arose because `Option` was mapped to a value both in the type
def and also in a type application in the test code. To fix this, I've
added a scope of top level definitions. When parsing arguments we check
against this dictionary to make sure we don't overwrite the arguments
for type functions defined elsewhere.

While fixing this bug, I noticed an issue with how we define the number
of arguments for a function. Instead of relying on the args specified by
the type definition, we relied on the variables counted in the function
body. This would lead to errors if we for whatever reason didn't use all
type arguments in the type body. I've added a test for this behavior.