Adding 'main' function and ability to run code from kind
To evaluate kind code, we need a way to pass some code and the arguments
we're wanting to run. In this commit I introduce the convention that a
`main` function which isn't exported in a module is called.

To do facilitate this, I've added a step to codegen to compile a main
module for any source which contains a main function. This main module
exports only the main function and nothing else.

In the `kind` module, we use this fact to compile the code given, and
call the main function afterwards before cleaning up the compiled
modules.