Add sandbox mode
When running `kind:run`, sandbox mode is enabled by default. This makes
it impossible to import any erlang module that I gives system access.
This is useful when compiling code of unknown origins, like what I do on
the kind website.

To compile the whitelist I had a look through erlang modules and tried
as best as I could to only include modules and functions that were
obviously safe.

The safety of this whitelist only goes so far. I'm hopefully that the
checks will discourage lazy vandalism, but if an actor is intent on
compromising the system in order to bring it down, I'm confident they'll
find a way. And if they succeed, well, then so be it until I bring it
back up again. Progress, not perfection.