Moved typegen tests to better testing structure
For reasons that I struggle to understand, rebar3 or eunit is not
picking up tests files in sub-directories any more. As far as I can
tell, this was the case when I started working on these changes a few
hours ago, but apperantly it's not the case any more, and I'm none the
wiser after a great deal of googling as to why this has happend.

I'm reluctant to spend any more time on this issue so I've moved all
tests files in to the test directory. It feels dumb and I don't like
that I don't understand what is happening, but outside of setting
`DEBUG=1` and noticing that the files aren't picked up I haven't managed
to make progress on tracking down this change in behavior.