Refactored typegen to use separate pattern gen lib
Pattern code generation is similar for types and defs. To not have to
write the code again I refactored out the part of typegen that dealt
with patterns.

In doing so I came across a few bugs and omissions. In particular the
support for qualified types (e.g. module imports) was rudimentary and
didn't check that the qualified type was callable with zero arguments.

When we call a type function at compile time to get a type domain, we
are returned a domain, rather than a term. Domains and Terms are a
source of confusion in the source code because they look similar,
particularly because I rewrite terms such as pairs in to `tagged` terms,
which is only different from a tagged domain in that the term carries a
`Ctx` containing for example the line and source file.

To write out erlang core code for the domains I need to either translate
them to terms or add support for domains in the AST traverser in
`ast.erl`. I'm reluctant to do the latter because the domain's lack of a
context map makes it impossible to point to source code in errors.

For this reason I've created a utility function that rudimentarily
translate domains to terms. I suspect that if we want to support type
application in patterns, we'll need a function that goes the other way
too.