Type sums are represented by a 'sum' term instead of a 'tuple'
When compiling code for a type, there's quite a bit of overlap with code
for a def. I'd like to combine code generation of both in to a single
module. This would (amongst other things) make it very simple to add
support for evaluating arbitrary expressions as part of a type
constructor. Or at least I think it will.