Fixed issue with non-type-application in typegen
Inside a type definition we encounter type applications when a type
constructor is applied to arguments, and normal application when a type
variable is applied to arguments.

For the latter case, the type variable needs to evaluate to a function
domain `{f, <Tag>, <Function>}` to be callable. In the previous
implementation it was assumed that the type variable would evaluate to
the type name instead. I've fixed this issue and added a test for it.