Fixed application of qualified types
Calling a qualified type only used to work if the type took no
arguments. This commit fixes this bug. Instead, when compiling a
qualified type, we want to call the domain function of the module where
the type is defined. In a type application this results in applying
arguments to a `{f, Tag, Func}` domain.

This solution works nicely for expressions, but in patterns erlang
doesn't allow function calls which means that we can't call `domain` at
runtime and will have to call it at compile-time instead if we want to
support patterns containing type constructors.

While fixing the issue of qualified types constructors with arguments I
came across a couple of import issues:
- Imports didn't support importing types with multiple non-type path
  components (e.g.  `kind/prelude/Option`). By calling `domain` instead of
  the type module,
- Domain functions for type modules wouldn't contain functions for the
  type module members, e.g. the domain function of the type module for
  `Boolean` didn't contain `True` and `False`.
- Similarly domain functions for ordinary modules didn't contain
  exported type members, e.g. the domain function of `kind/prelude` didn't
  contain functions for `True` and `False` either.