Simplified typegen + debugger .state file
When generating code for a type, we can just call the domain function
for the type instead of traversing the domain and generate code for it.

What might be even simpler (but haven't done it yet) is to include a
transformation so that when an already created type `T` is referred to, we
can rewrite the code as `T.domain` and always return the atom
corresponding to the type.

I've added a .state file for debugging. The paths in the file are
absolute so it's of absolutely no use for anybody but myself. To debug
using the file run:

```
rebar3 as test shell
> debugger:start(".state"). // And add break-points
> r3:do("eunit --generator <module>:<test name>").
```