Expr in Application Term can be both type domain or function
To apply a type function to arguments, I need to extract the function
from the `{f, Tag, <Function>}` domain first and then apply it to
arguments. For a normal function we can do this straight away.

Without knowing the domain of the function we can't decide at compile
time if the Expr evaluates to a type constructor or a normal function.
Instead we currently have to pattern match at runtime. That's a bit
naff, but hopefully it's trivially fixable once I've re-implemented the
scanner to embed domains in the Ctxes of the AST.