Added support for strings, numbers and atoms in one fell swoop
Strings are represented as utf8 encoded binaries (similar to Elixir).

Pattern matching binaries turned out to be a faff. To create a
reference, I transpiled some erlang code to core and inspected it after
parsing it in the interpreter. This might be useful some other time, so
I'm just going to jut down a quick how to in this commit for future
reference:

```
echo "-module(some_mod).\n-export([f/1]).\nf(<<\"❤️\"/utf8>>) -> true;\nf(_) -> false." > some_mod.erl
erlc +to_core some_mod.erl
erl
> {ok, FileContents} = file:read_file("some_mod.core").
> Core = unicode:characters_to_list(FileContents, utf8).
> {ok, Scanned, _} = core_scan:string(Core).
> io:format("~tp~n", [core_parse:parse(Scanned)]).
```

This gave me the core binary representation of the heart emoji as
represented in a pattern:

```
{c_binary,[],
    [{c_bitstr,[],
         {c_literal,[],10084},
         {c_literal,[],undefined},
         {c_literal,[],undefined},
         {c_literal,[],utf8},
         {c_literal,[],[unsigned,big]}},
     {c_bitstr,[],
         {c_literal,[],65039},
         {c_literal,[],undefined},
         {c_literal,[],undefined},
         {c_literal,[],utf8},
         {c_literal,[],[unsigned,big]}}]}
```

From which I learned that:

* Every unicode code-point (Heart & Red) must be wrapped in `c_bitstr`
* All terms must be wrapped in `literal` (even the code-point value)
* `c_bitstr` for `utf8` doesn't need to specify `Size` and `Type`

Fun times...

inspect the output
of erlang transpiled to