Changed product domain from {product, M} to just M
Domains come in two types: Data values like List and Map that currently
map directly on to data structures and domain values like Sum and Tagged
which don't map on to erlang data types and tend to describe a domain
rather than a data type.

Tagged is a bit of a half-way house. The way pattern matching works
currently, a value has to be wrapped in `{tagged, tag, V}`. I'm hoping
that once I know the domain of values, I can insert filter the patterns
in a pattern match that aren't tagged with the same tag out and not
generate code for them at all (except maybe in types).