Fixed two type gen issues
Issue 1: Local unexported types were not callable from within the same
file, and no domain function would be compiled if the file had no module
declaration. The fix was to generate both exported and non-exported
types along side the domain function and pass them to `def_gen`.

However, with this fix comes a different issue of limiting the types we
export in `def_gen` to only the types exported by a module. This isn't
done yet.

Issue 2: When defining a pattern for a type `Test -> Option(Boolean)`,
`pattern_gen` would error because compiling the domain of test meant
compiling a type application (`Option(Boolean)`) which wasn't supported.
To fix this issue I've added support for qualified type application.

There still remains to be support for general type application although
it should be trivial at this point. There's also a lack of tests which
need to be added.