Fixed grammar for pattern applications
I had some idea that I would add local type applications as well, but
that turned out to be harder than I had given it credit for.

Take a pattern like `F(A)`. Since `F` can contain arbitrary function
calls, we have to evaluate `F(a)` in order to know what it might return,
but since it's defined in the same file, it might call the very type
that we're compiling.

Normally in code we handle this by compiling to a function call which
will then happen at runtime. This is also what we do in a type
expression (e.g. not a pattern), but in a pattern, erlang core doesn't
accept `cerl:c_apply` or `cerl:c_call` as valid forms.

It's possible to call type constructors from other modules in the
pattern, but local ones will lead to a compile time error.