Infix notation for operators
This was a tricky change and has taken up pretty much the last three
days to get right. Getting infix operators to work means dealing with
right and left associativity (whether an operator is grouped from the
left or the right, similar to a left or right fold). It also means
dealing with priority (so that multiplication happens before addition
for example).

As part of this change, I've:
- Refactored qualified_type and variable to qualified_symbol: Infix operators
  can be both types and variables and can't be told apart just based on
  the symbol used. Instead of differentiating type operators and variable
  operators I got rid of the distuingtion between qualified types and
  qualified symbols.
- Introduced `operator` and infix application in grammar: It was a royal
  pain to incorporate infix application in the grammar, since it was
  fairly ambigious already. To aviod ambiguity, I've had to redefine the
  syntax for lambdas to always involve a vertical bar. I'm not sure how I
  feel about this change, but I couldn't see a way around it.
- Set operator precedence according to first character in symbol: It's
  possible to define any operator one might want, but I don't have a
  syntax for setting the operator precedence like for example Haskell
  because of how `yecc` works. Instead I have defined operator precedence
  (and left/right association) based on the first character of the
  operator. To give an example, an operator named `*=` would have
  precedence over an operator named `+-->` because `*` has precedence over
  `+`.
- Fixed so that type dependencies are no longer filtered out.

Precedence
----------

The current operator precedence (listed from highest to lowest):
  ^	(exponentiation)
  *	(multiplication)
  //	(division)
  +	(addition)
  -	(subtraction)
  <	(less than)
  >	(greater than)
  =	(equals)

Right-associativity
-------------------

I've been fighting with yecc and not found a way to get both precedence
and association of operators completely right. For now I've settled on a
compromise of sorts. A few operators are right associative, amongst them
division which really should be left associative. I had the choice
between giving division the correct precedence but right associative or
left associative but wrong precedence. This felt like the better choice.

Any operator that starts with a colon (`:`) will have right
associativity (similar to scala).

To fix:
- Lambdas don't have an end. That means that when they aren't wrapped in
  brackets, they keep parsing and err when the next line isn't a lambda
  clause. There's failing tests that I suspect are because of this
  currently
- Associativity (see commented out test in syntax_test)
  - Tagger needs to collect operators when they are the key in a pair as
    types
  - Fixed bug in import module (dependencies of type dependencies were
    filtered out)
  - Fixed right and left associativity to wider infix application terms