Added macro support
Macros rewrite the internal abstract syntax tree at compile time. They
work like normal functions and share their syntax, but are created with
the `macro` keyword rather than the `def` keyword.

I'm introducing macros as a step on the way to be able to evaluate
arbitrary domains. For an arbitrary expression, I want to be able to
call `.domain` and have it return the domain of the expression.

Next steps are:
- Create AST module in kind covering the terms used for the syntax tree
- Scan top-level definitions and create a function which given a
  function name returns the domain
- Create domain macro in kind which transforms syntax tree to domain