Function parameters can be patterns
This is the visible effect of a pretty thorough simplification of how
functions are interpreted. Instead of having `def` and `types` terms
contain args and expression, they now contain a function. This matches
them more closely with how erlang core functions are represented, but it
also simplifies the code, both conceptually and in terms of verboseness
because a `def` now simply maps a name to an anonymous function, and
that function is generated the same whether it's a type, a def or a
macro for that sake.

Changes include:
- Simplification of syntax.erl: All defs/types/macros matches a pattern.
No separate `def_clauses` or `type_clauses`
- Change in syntax for pattern matching. We no longer list the variables
of the function if they are pattern matched. If we need a top level
variable and want to match on a pattern we can use the `(var: pattern)`
syntax.
- Change in type sums: Because types aren't separate from other code,
type sums need to be wrapped in brackets to not be ambigious. I'm
wondering if I can replace them with a function all-together.