1/?: Rewrite source modules early on to match generated erlang modules
In erlang a source file maps on to a module 1 to 1, but in Kind this
isn't the case because a source file can export several modules and type
members.

So far, we've tackled this at the end of the compilation pipeline, by
making sure that once we've generated the erlang core code, we create a
labyrinth of modules with functions calling each other. See `domain_gen`
and `def_gen` for examples. To do this at the very end exposes the
complexity of a module-mapping separate from he AST to the rest of the
compiler.

With this change, one of the first things the compiler does is to parse
the source file in to it's constituent modules. These modules can then
be typed and compiled individually.

This change is incomplete in that it only encompasses the changes to the
`modules.erl` component. The next step is to

Module Translations
-------------------

The translation from source file to modules takes place as follows:

- Each source file can define multiple modules. The definitions in the
source file can either be inside a module and only available to that
module. They can also be defined outside the modules and are available
for all modules.
- All definitions outside modules are gathered in a root module which
exports all members. Any module that exports a function defined in the
root module, is equipped with its own function definition with a call to
the root module function.
- Each type defines a module of all its sub types. The sub modules don't
contain the original type definition, but only the sub types defined
inside the type defition. For each subtype that isn't just a
value-literal, we define a constructor which allows us to instantitate
the type value.

Module Statements
-----------------

In addition to the above changes (I know...) I've added syntactic
support for statements (e.g. definitions and imports) within a module.
This has several usages:

- A source can define e.g. a prod and test module that defines a few
constants and otherwise exports the same definitions from the source
- In testing I can test the interactions of different modules by
defining multiple modules in the same source instead of having to make
my test cases support multiple source files (which is a faff)
- Online it makes it easy to have a single source text field that
showcases several modules

Type Constructors
-----------------

The definitions for subtypes are replaced with type constructors using
the `tagged_gen` module. I'm a bit worried that the AST Terms generated
by this conversion reflect a later compiler pipeline step. We'll find
out I guess.

Performance
-----------

Having a lot of functions in different erlang modules is most certainly
not a performant way of writing code. For now though, I don't think it's
very important. I'm sure it can always be changed later.