Reworked import module to support new module structure
Changes include:

- Support for type modules instead of passing in local types.
- Support for import statements in modules. Any import in the root
modules also apply to modules defined in the root module.
- Support for importing individual local types. Due to an omission, the
import module didn't support local imports of the type `T/A` before.
- Improved tests with support for providing a module map by writing Kind
code.
- Support for transitive local imports. This means that we can import
types from an external module as if they are locally defined types:

```
module a {T} (type T -> (A | B))
module b {} (import a/_
             import T/_)
```

This latter change is important because it means that type members defined in
the root module can be imported in the sub modules with the same syntax
as one would have used to import them in the root module.

Future Changes:

What I'm noticing, is a module and import grammar using `/` which is
half-supported. I can call a module `a/b/c`, but I can't import `a/b`
and then refer to `b/c/...` in the code. Currently the import system
only supports me importing the full module name `a/b/c`.

With transitive local imports I added in support for partial paths (like
`a/b` in the example above), but the support is patchy when a partial
module path can't currently be imported itself.

I think it would be nice, but possibly not wildly important to rewrite
the import module to work with partial path support from the ground up,
instead of the current translations to and from beam module names.