Added linked definitions
There are a few different circumstances when a module contains a
definition which is defined elsewhere. In this commit I introduce the
`link` term and use it in the following use cases:

- Exported definitions: When a module is exporting defs that are defined
  in the root module, each def is added to the module as a link pointing
  to the root module
- Borrowed constants: When a def refers to a constant defined elsewhere,
  the module for the def contains a link pointing to the definition of the
  constant

The usage of links makes it simpler to handle references to constants
and defs that don't point to where they are defined. Take as an example
the following bit of code:

```
def T -> (A | B)
import T/B
def S -> (B | C)
```

Because `T/B` is imported in to the local scope as `B`, any reference to
`S/B` elsewhere in the code should translate to `T/B`. This property
makes it seamless to gather tags and constants from other definitions
under a new constructor.

Implementation
--------------

In the `module` module, links are created for defs and constants where
appropriate. Care is taken to ensure that constants that are imported
several times over refer to the original def where they are first
defined.

In the `import` module, any imported def is checked to see if it is a
`link` term. In this case a `qualified_symbol` pointing to the link
destination is used instead.

In the `tagger` module, links form part of the local scope like any def

Notes
-----

A lot of small bits and pieces had to be put in the right place for
links to be implemented. Here is a list of particular points of
interest:

- Simplified import terms: In `module` I sanitise all import terms so
  they don't contain the parsed symbols, but just the extracted values.
  This simplifies both `module` and `import` and their tests in
  particular. As a consequence a lot of reshuffling of `import` has
  happened
- Recursively resolving imports: In `module` I need to walk the trees of
  possible imports being used by other defs and then imported again under
  a new name. This led to a little algorithm I'm quite pleased with under
  the name `expand_imports`. It makes sure that any imports under a
  different name are traced back to the original definition and name of
  a constant to make sure the links point to the right place.
- Tagged Sources: In the parser I've added an option for sources to come
  in `{text, tag, Code}` triplets where the tag will be used to refer back
  to the source (similar to a filename for a path). This makes it easier
  to use in testing because tagged sources won't be indexed under randomly
  generated names.
- Imported defs are removed in sub modules: When a submodule defines a
  def which is also an imported value (in particular when said def is a
  link), the def is replaced by the imported value. I can't remember why
  this is useful.
- Parser tests are passing. They needed a bit of love.