Fixed dependencies between local types and source module
A source module would depend on local type modules defined in the source
module, but these modules would in turn depend on the source module for
any functions called in their definitions.

When I coded up this design, I thought it wouldn't be an issue since I
could just make sure I wouldn't catch the cyclical dependencies when
running the parser. However, it turns out that having cyclical
dependencies was a bad thing after all.

When I create domain modules for the type checker, I make calls to
domain functions for all module dependencies. For cyclical dependencies,
these domain modules haven't been defined and loaded and the calls fail.

I've fixed this issue by moving all the actual function definitions for
type members in to the source module and then link to them from the type
modules. This maintains type module transparency. When importing
`Boolean/True`, there will be a `Boolean` module defined, containing and
exporting a definition called `True`. However `True` will link to the
source module where we also define a function called `Boolean/True`.

Local Imports
-------------

Local imports end up pointing back to the module where they are made
when sub defs are no longer in a separate module specifically for the
type. This leads to a circular dependency where the symbols are replaced
by a qualified symbol pointing to the same module. When this is
evaluated during type checking, we'll end upn trying to make a function
call to a domain function a module while it's under construction. No
bueno.

To fix this, I am replacing the qualified symbols with a type term. This
isn't what I have in mind for the long term, but for now it keeps the
code in line with past behavior. This has resulted in changes in the
`tagger` module and in the `import` module dealing with qualified
symbols pointing back to the same module. It has also meant changes in
the `module` module in order to move sub defs back to modules and link
to them from sub modules.

Normalized Applications
-----------------------

A qualified symbol is due to be either the expression part of an application,
or itself an application with zero arguments. To make this simpler,
qualified symbols are replaced with qualified applications after
parsing.

Symbol Path
-----------

We occasionally store a path as an atom using the '/' separater between
path elements. This is true for example for sub defitions in a module
like `type T -> A` where `A` can be accessed under the def `T/A`. On the
flipside we sometimes need to break this construction apart in to its
path. And really, because this is necessary, it hints at the
representation being shite, but refactoring it is a big and not very
rewarding task, so instead I've put this workaround in place.

Other things to note
--------------------

- I've redone a bunch of tests so I test every component of a def-map or
a module-map separately. That makes it easier to spot where a problem is
when a test is failing.