Homogenised import module
I was finding myself adding edge case after edge case in the import
module to support a consistent grammer of imports, but for each edge
case I added, the code would get more and more convoluted and hard to
follow. While it maybe wasn't a high priority, it felt like a fun
challenge to rephrase the code to deal with the grammer of imports in a
consistent manner.

The grammer of imports is broad strokes the following:
- Import statements in a source file or module are understood in the
  order they appear. Later statements can refer to modules imported from
  earlier statements.
- Module names are defined as a path with the path elements separated by
  forward slashes ('/'). It's possible to import part of this path. To
  give an example, consider the following module:

  ```
  module a/b/c {f, g} (def f -> noop
                       def g -> noop)
  ```

  The import statement `import a/b` would bring `b/c/f` and `b/c/g` in
  to scope.
- If in addition to the import statement `import a/b`, there was a later
  import statement of `import b/c/f` then `f` from the `a/b/c` module
  defined above is brought in to scope.

Before this change, the above would be possible using local imports, but
not with imports from another module. This change both generalises the
grammar and also cleans up the import module substantially.

Smaller changes as part of this commit:
- Introduced new AST terms `beam_symbol` and `beam_application`. This is
  done to more easily treat calls to compiled beam modules separately from
  calls to kind modules.
- The import module takes a module as a parameter instead of a
  particular import. This moves the code for checking duplicates in to the
  import module and simplifies the parsing
- Fixed error in `module` module where links containing qualified
  symbols were broken because the module path hadn't been cleaned from
  {symbol} noise