Removing type defs and introducing keywords instead
Type defs and defs encode the same behavior, however type defs were
treated differently and parsed in to `type` terms together with the
keywords and tagged functions defined inside them.

This `type` term added needless complexity down the compiler chain
because it covered over several disparate concepts (a constructor, a
keyword, a function). Conceptually the existance of two more or less
identical definition types makes the language syntax harder to reason
about.

As a consequence, I've taken out the type def cosntructors and `type`
terms. They are replaced as following:

 - Type Defitinitions -> Normal Definitions: Any def can represent a
   "type" constructor. The distringuishing feature between running a
   constructor to evaluate a domain and running a function to evaluate
   code is the context where the function is called. If called in a
   pattern, the function is evaluated in order to compute a domain at
   compile time. If evaluated in an expression, the function is evaluated
   to compute a value at run time.
 - Type constants -> Keywords: Upper case symbols that aren't in the
   import scope are defined as a keyword pertaining to the function. A
   keyword when imported is imported as a keyword value rather than being
   a qualified symbol. This clearly ambiguates the cases where a call to a
   function is needed from the cases where a keyword can be used as is.
 - Tagged Defs -> Sub Definitions: When a domain is tagged by a keyword,
   a constructor is created in order to make it possible to construct
   values of this type. These values are boxed by the keyword at compile
   time making sure that a boxed value like `Dollar: Int` can't be
   replaced by another boxed value like `Pound: Int`. At runtime the
   values are not boxed.

This change covers the introduction of keywords and the change from type
defs to defs for the parsing stage. Next stage is to complete the module
transformation for the type checking and compilation stage. This step
should hopefully be simplified by the type terms no longer existing.

Other changes:
--------------

 - Calls to compiled erlang/beam modules have their own separate term:
   `beam_symbol` and `beam_application`.
 - Sub modules only export types that are exported by the parent module.
   They no longer need to expose all type members, since they only link to
   the type definition in the parent module anyway.
 - Tagger no longer needs to look up type symbol
   based on it's parent scope. The only points where a `{symbol, _, type,
   _}` term exists, is in the original definition at which point it can
   trivially be looked up in the scope and converted to a keyword.