Correctly propagate pattern envs
Take the following def clause:

```
def f (a: 1 | 2) -> a
```

If `f` is called with an argument which at compile time has a domain of
`1`, then `f` will have a domain of `1`. However, when `f` is called
with `any`, then the sum part of the pattern would be expanded to two
clause lines and the function would end up rendered like so:

```
def f (a: 1) -> 2
      (a: 2) -> 2
```

As is obvious, the result should not be `2` in both lines. This
happened, because we only computed the clause expression (`a` in this
case) once per clause with the union of environments (e.g.
`maps.merge({a: 1}, {a: 2})`).

In order to separately compute the expression with the environment
specific to the particular sum expansion of the clause, I found myself
in a pickle. To expand sums, we let the `pattern` function return a list
of patterns, and then in the `clause` function we create a list of
clauses accordingly. However, using the `ast:traverse` function, we
still only return a single environment, and so all environment values in
the list of patterns get merged and cause the problem described above.

To circumvent this behavior, I ended up storing the pattern environment
as part of the pattern and propagating it separately. This is a bit
awful in that it introduces quite a lot of complexity in the code, but I
could not think of any more straightforward ways of solving this issue.
I suspect that if I could separate out the pattern sum expansions from
the linearization, it might be neater, but that's a bigger and more
invasive change, so I was reluctant to undertake that work.

One of the subtelties of passing the environment along as part of the
pattern terms, is that for expressions like `pair` where the pattern is
computed separately, we need to be mindful of propagating the `env`
key/val of the context.

I've introduced two unit tests covering these use cases.