Use qualified path for tagged terms
This commit is part bug-fix, part change.

As a bugfix, the commit addresses an issue where tagged types mentioned
in one def but imported into scope from another definition will use the
path of the current def and not the correct path where they are defined.

This issue stems from the path of a tagged term being set in the
`module` module and not in the `tagger` module. At there point where it
was set, there was no access to scoping information for imported defs
defining similarly named tagged terms.

As a fix, I've moved the responsibility for setting the correct path in
a tagged term to the `tagger` module according to the import scope. At
this point it is trivial to fully qualify the path, and I've chosen to
do so because it otherwise would be ambigious if two modules had defined
a tagged type with the same constructor and keyword.