Introducing domain normalization
In the past I've used a function `compact` to in essence compact the
representation of a domain so that e.g. `A | (B | C)` is flattened to
`A | B | C` and `{a: 1} | {a: 2}` is flattened to `{a: 1 | 2}`. However,
at the linearization stage we rely on any sums in a returned domain to
 a) Not clearly defined making it harder to make domain code that works
    for all cases
 b) Made working with sums in the linearization stage more difficult

What I mean by not clearly defined is that outside of looking more
compact and less redundant, the `compact` function didn't follow any
clear rule in order to arrive at an unambigious canonical representation
of a domain. I don't have examples at hand, but it's entirely plausable
that two different representations of the same domain might have ended
up compacted differently. This in turn makes it hard to produce
functions that work on domains, because they will have to work for all
possible compacted domain shapes.

At the linearization stage I need domains to always be composed such
that any sums have been propagated to the outer-most value in the tree
representation of the domain (so exactly the opposite of what `compact`
was doing). As I wrote the code with this assumption, I realised that
this particular representation of a domain is analoguous to a boolean
disjunctive normal form [1].

And so in order to simplify the domain logic, I've replaced the
`compact` function with a function that computes the disjunctive normal
form of a domain. Because `domain:normalize` is always called when using
`union` or `intersection`, it means that the domains returned from the
`domain` module will always be in disjunctive normal form.

[1]: https://mathworld.wolfram.com/DisjunctiveNormalForm.html